invoice look up вынесен в отдельный класс
